@@ -40,13 +40,13 @@ namespace ASC.Common.Tests.Web
         public void GetExtentionTest()
         {
             var ext = MimeMapping.GetExtention("application/x-zip-compressed");
-            Assert.AreEqual(".zip", ext);
+            Assert.Equals(".zip", ext);
 
             ext = MimeMapping.GetExtention("Application/x-zip-Compressed");
-            Assert.AreEqual(".zip", ext);
+            Assert.Equals(".zip", ext);
 
             ext = MimeMapping.GetExtention("Application/ZIP");
-            Assert.AreEqual(".zip", ext);
+            Assert.Equals(".zip", ext);
         }
     }
 }
